refactor: correct spelling error
